Edit 1

+++ app/Http/Controllers/API/AdminAPIController.php
@@ -15,40 +15,35 @@ class AdminAPIController extends APIController
     // Get all pending users
     public function pendingUsers()
     {
-        // if (auth()->user()->role !== 'admin') {
-        //     return response()->json(['error' => 'Unauthorized'], 403);
-        // }
+        if (auth()->user()->role !== 'admin') {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
         
         $pending = User::where('is_approved', false)->get();
-        // return response()->json($pending);
-        return UserResource::collection($pending);
+        return UserResource::collection($pending) ->additional([ 'message'=> 'Pending users:']); 
 
     }
 
     // Approve a specific user
     public function approveUser($id)
     {
-        // if (auth()->user()->role !== 'admin') {
-        //     return response()->json(['error' => 'Unauthorized'], 403);
-        // }
+        if (auth()->user()->role !== 'admin') {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
         
         $user = User::findOrFail($id);
         $user->is_approved = true;
         $user->save();
-        return new UserResource($user);
-
-
-        // return response()->json(['message' => 'User approved successfully']);
+        return (new UserResource($user)) ->additional([ 'message'=> 'User approved successfully']);
     }
 
     // Reject (delete) a specific user
     public function rejectUser($id)
     {
-       // if (auth()->user()->role !== 'admin') {
-        //     return response()->json(['error' => 'Unauthorized'], 403);
-        // }
- 
-        
+       if (auth()->user()->role !== 'admin') {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
+
         $user = User::findOrFail($id);
         $user->delete();
         return response()->json(['message' => 'User rejected and deleted'], 200);
